+++ uncopy_me/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-# Copyright (C) 2020  Marcus Rickert
+# Copyright (C) 2020-2022  Marcus Rickert
 #
 # See https://github.com/marcus67/uncopy_me
 # This program is free software; you can redistribute it and/or modify
@@ -14,3 +14,4 @@
 # You should have received a copy of the GNU General Public License along
 # with this program; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+