+++ CHANGES.md
@@ -4,6 +4,13 @@
 
 This document lists all changes of `UncopyMe` with the most recent changes at the top.
 
+## Version 0.1.2 (January 16th, 2022)
+
+* Add several test cases
+* Add test setup mode (option `--use-test-setup`)
+* Add option `--force` for prompting before deletion
+* Add interactive prompting class `BasePrompter` 
+
 ## Version 0.1.1 (January 16th, 2022)
 
 * Initial version